Refactor TMS architecture by updating namespaces to use Infrastructure layer and adding new interfaces for VaultClient and SQL session management
