+++ TMS.Infrastructure/Extensions/ServiceCollectionExtension.cs
@@ -2,9 +2,9 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using TMS.Core.Extensions;
 using TMS.Core.Interfaces.Factories;
-using TMS.Core.Interfaces.HttpClients;
-using TMS.Core.Interfaces.Persistence;
-using TMS.Core.Interfaces.Persistence.Procedures;
+using TMS.Core.Interfaces.Infrastructure.Http;
+using TMS.Core.Interfaces.Infrastructure.Persistence;
+using TMS.Core.Interfaces.Infrastructure.Persistence.Procedures;
 using TMS.Infrastructure.Factories;
 using TMS.Infrastructure.Http;
 using TMS.Infrastructure.Persistence;
