+++ TMS.Infrastructure/TMS.Infrastructure.csproj
@@ -3,6 +3,7 @@
   <ItemGroup>
     <PackageReference Include="Dapper" />
     <PackageReference Include="Microsoft.Data.SqlClient" />
+    <PackageReference Include="Microsoft.Extensions.Http"/>
   </ItemGroup>
 
   <ItemGroup>
